Refactored TrafficMonitorBase to allow for separaration of monitor-specific arguments
This improves separation of concerns and will allow for code to be selectively included. e.g. if a participant wants to use just one type of capture monitor, the code for the other monitors can be completely dead/disabled - reducing the amount of code to audit.

Only a stub implementation of the DIS-FAST upload support is in this version.